Create async version and generate sync version from it
